requirements.txt
argparse
pysmbclient
